manabar

+++ Skaven-Warren-Defense/Assets/Scripts/Tony's Scripts/healthSlider.cs
@@ -9,13 +9,13 @@ public class healthSlider : MonoBehaviour
     // Start is called before the first frame update
     public Slider slider;
     
-    public void setMax(int health)
+    public void setMax(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
     // Update is called once per frame
-    public void CurrentHealth(int health)
+    public void CurrentHealth(float health)
     {
         slider.value = health;
     }
